1) Fixed dotenv package issues 2) Refactored app.ts startup 3) Upgrade module and typescript versions
