+++ package.json
@@ -17,32 +17,30 @@
     "test:circleci:summoner": "jest --testPathPattern=src/__test__/Unit/summoner.test.ts"
   },
   "dependencies": {
-    "@jest/transform": "^29.3.1",
-    "axios": "^1.1.3",
+    "axios": "^1.2.0",
     "body-parser": "^1.19.2",
     "cors": "^2.8.5",
     "dotenv": "^16.0.0",
     "express": "^4.17.3",
-    "mongoose": "^6.2.4",
-    "node-cron": "^3.0.0",
-    "npm": "^9.1.1",
+    "mongoose": "^6.7",
+    "npm": "^9.1.3",
     "swagger-jsdoc": "^6.2.5",
     "swagger-ui-express": "^4.5.0",
-    "winston": "^3.8.1",
     "ts-node": "^10.6.0",
+    "winston": "^3.8.1",
     "winston-loggly-bulk": "^3.2.1"
   },
   "devDependencies": {
+    "@types/jest": "^29.2.3",
+    "@jest/transform": "^29.3.1",
     "@types/express": "^4.17.13",
-    "@types/jest": "^29.2.2",
-    "@types/node": "^18.11.9",
+    "@types/node": "^18.11.10",
     "@types/supertest": "^2.0.12",
     "@types/swagger-jsdoc": "^6.0.1",
     "@types/swagger-ui": "^3.52.0",
     "@types/swagger-ui-express": "^4.1.3",
     "jest": "^29.3.1",
-    "nodemon": "^2.0.15",
-    "supertest": "^6.2.2",
+    "supertest": "^6.3.2",
     "ts-jest": "^29.0.3",
     "typescript": "^4.9.3"
   }
